large update
added js code to generate an list 'li" of pokemon that includes a button for each pokemon. created a function to add this list inside of the IIFE.

The index file now has a <ul class= 'pokemon-list'></ul> . this is to utilize the code in the .js file

added code to css "copied from lession" and made small changes for button appearance.